chore(nfts): remove utils file
